e2e: add a test for using private registry with podman driver
This PR adds an e2e test case that stands up a private docker registry
and has a podman task run a container from an image in that private
registry. Currently the podman driver only supports basic user:pass
authentication embeded in the task. We can use this registry to iterate
on future support for credentials helpers, etc.